Fix Oclock formatting

+++ types.go
@@ -127,8 +127,8 @@ func (b Bearing) Degrees() float64 {
 	return float64(360.0 * b / (2 * math.Pi))
 }
 
-func (b Bearing) OClock() int {
-	return int(b.Degrees() * 12. / 360.)
+func (b Bearing) OClock() float64 {
+	return b.Degrees() * 12. / 360.
 }
 func (b Bearing) Radians() float64 {
 	return float64(b)
